requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-redis==4.11.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.7.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
legacy-cgi==2.6.3
MarkupSafe==3.0.2
mongo==0.2.0
openapi-codec==1.3.2
packaging==25.0
pymongo==3.12.3
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.0
setuptools==80.9.0
simplejson==3.20.1
sqlparse==0.2.4
uritemplate==4.2.0
urllib3==2.5.0